src: use non-deprecated overload of V8::SetFlagsFromString

+++ src/node_v8.cc
@@ -139,7 +139,7 @@ void UpdateHeapCodeStatisticsArrayBuffer(
 void SetFlagsFromString(const FunctionCallbackInfo<Value>& args) {
   CHECK(args[0]->IsString());
   String::Utf8Value flags(args.GetIsolate(), args[0]);
-  V8::SetFlagsFromString(*flags, flags.length());
+  V8::SetFlagsFromString(*flags, static_cast<size_t>(flags.length()));
 }
